Adding a Background
For some reason I had to restart because nothing was saved properly. Because of this I had to restart my progress to make this game so far. Please note that I ealier today was when I found out about this problem but I didn't write about it in my description.