I tidied my files again to make clear space I also made white space consistant between files and changed the radius of the collision shape 2d node.
I didn't delete the other fire pngs in the folder flame 10 because I was unsure if I wanted to animate the fire spell and decided to put that decision to if I have time in sprint three. And I didn't mention this last time I deleted some files but I kept the attacking, hurt and idle animations because I have plans to use them this sprint.