I added a second wave.
I added a second wave. I also moved where Weak Ghost and Weak Ghost 3 are positioned as I thought they were taking too long to get on screen after I fiddled with the animations to spawn as for some reason you could see the ghost head to where they spawn at the top of the screen. This was finished on Friday the 28th.