Added a water spell that slows the mobs
There was a big problem with getting it to shoot properly as it followed the rotation of the spellbook so I made the book rotate around the pivot point not the origin point which I thought was happing but was obviously not. Because I added a new timer for when the effect would wear off I renamed the timer nodes to death as it would be the death timer node to be called upon during the dying animation.